# server-app/api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kanban application",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:9900/api"
        }
    ],
    "paths": {
...
        "/auth/user/login": {
            "post": {
                "tags": [
                    "user"
                ],
                "operationId": "userLogin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "title": "UserLoginRequest",
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "title": "UserLoginResponse",
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/UserDTO"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserDTO": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "username": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "avatarUrl": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}